Available Stan Probability Distributions

A data.frame listing all the Stan probability distribution functions available within this package.

available_stan_probability_distributions

Format

A data frame with 1 row and 5 variables:

function_name

Name of the function

builtin

Whether the function is built-in to Stan

distibution

Which distribution the function relates to

function_type

Type of function, e.g. PDF or Quantile

log

Whether the function returns log values

Details

The source-code for a function can be loaded using the stan_probability_distribution function

Examples

available_stan_probability_distributions
#>             distibution function_type  log builtin   function_name
#> 1 zero-inflated Poisson           PMF TRUE   FALSE zi_poisson_lpmf